refactor: extract useUploadFormState hook from UploadForm
Moves the 24 useState declarations and the resetForm function out of
UploadForm.tsx and into a colocated hook. Pure refactor — behavior is
identical, verified by the existing 27 UploadForm tests passing
untouched. UploadForm.tsx is now ~80 LOC smaller and new state added
to the hook auto-handles the reset.
